+++ src/components/DistrictBarchart.svelte
@@ -1,30 +1,30 @@
 <script>
     import { scaleBand, scaleLinear } from 'd3-scale';
     import { slide } from 'svelte/transition';
-    
-    
-    
-    export let total_count, district_data, district_data_filtered;
+
+
+
+    export let total_count, districts_data, districts_data_filtered;
 
     let xDomain, xScale, yScale;
     const dimns = { width: 800, height: 300, label_x: 180, gap_x: 10 };
 
-    $: if (total_count != 0 && district_data) {
-        district_data.sort((a, b) => b[1] - a[1]);
-        district_data_filtered = district_data.slice(0, 5);
-        
-        xDomain = district_data_filtered.map((d) => d[1]);
+    $: if (total_count != 0 && districts_data) {
+        districts_data.sort((a, b) => b[1] - a[1]);
+        districts_data_filtered = districts_data.slice(0, 5);
+
+        xDomain = districts_data_filtered.map((d) => d[1]);
         // Prevent domain collapsing to midpoint if all values are zero
         if (Math.max(...xDomain) == 0) {
             xDomain.push(1);
         }
-        
+
         xScale = scaleLinear()
             .domain([0, Math.max.apply(null, xDomain)])
             .range([0, dimns.width - dimns.label_x - dimns.gap_x]);
-            
+
         yScale = scaleBand()
-            .domain(district_data_filtered.map((d) => d[0]))
+            .domain(districts_data_filtered.map((d) => d[0]))
             .range([0, dimns.height])
             .paddingInner(0.5);
     }
@@ -33,9 +33,9 @@
 
 
 <svg class="w-full h-auto fill-black" viewBox={`0 0 ${dimns.width} ${dimns.height}`}>
-    {#if district_data_filtered}
+    {#if districts_data_filtered}
         <g class="fill-red">
-            {#each district_data_filtered as d}
+            {#each districts_data_filtered as d}
                 <text
                     text-anchor="start"
                     class={d[1] == 0 ? 'invisible' : 'text-xl fill-black'}