+++ src/components/FireCountChart.svelte
@@ -1,38 +1,43 @@
 <script>
     import { scaleLinear, scaleTime } from 'd3-scale';
-    import { axisLeft, axisRight, axisBottom } from 'd3-axis';
+    import { axisRight, axisBottom } from 'd3-axis';
     import { max, extent } from 'd3-array';
     import { timeMonday } from 'd3-time';
     import { timeFormat } from 'd3-time-format';
+    import { format } from 'd3-format';
     import { select } from 'd3-selection';
-    
-    export let data;
-    
+
+
+
+    export let data, height = 300;
+
     let svg;
-    const dimensions = {
+    let dimensions = {
         width: 600,
-        height: 300,
-        marginLeft: 40,
-        marginRight: 40,
+        height: height,
+        marginLeft: 20,
+        marginRight: 36,
         marginTop: 10,
         marginBottom: 30,
     };
-    
+
     $: if (data) {
-        data = Object.keys(data).map((key) => [key, data[key]]);
-        let data_filtered = data.slice(data.length - 30)
+        let data_tuple = Object.keys(data).map((key) => [key, data[key]]);
+        let data_filtered = data_tuple.slice(data_tuple.length - 30)
             .map((data) => [new Date(data[0]).setHours(0, 0, 0, 0), data[1]]);
-        
+
         // Select the svg element
         let svgSelection = select(svg);
-        
+
         let xScale = scaleTime()
             .domain(extent(data_filtered, (d) => new Date(d[0])))
             .range([dimensions.marginLeft, dimensions.width - dimensions.marginRight]);
+        let yMax = max(data_filtered, (d) => d[1]);
+        // Prevent 0 tick to show up in the middle of y axis and add two extra ticks if all values are zero
         let yScale = scaleLinear()
-            .domain([0, max(data_filtered, (d) => d[1])]).nice()
+            .domain([0, yMax + Math.max(Math.ceil(0.4 * yMax), 2)]).nice()
             .range([dimensions.height - dimensions.marginBottom, dimensions.marginTop]);
-            
+
         // Add the x-axis and labels
         svgSelection.append('g')
             .attr('transform', `translate(0, ${dimensions.height - dimensions.marginBottom})`)
@@ -40,11 +45,11 @@
         // Add the y-axis and labels
         svgSelection.append('g')
             .attr('transform', `translate(${dimensions.width - dimensions.marginRight}, 0)`)
-            .call(axisRight(yScale).tickSizeOuter(0));
-        svgSelection.append('g')
-            .attr('transform', `translate(${dimensions.marginLeft}, 0)`)
-            .call(axisLeft(yScale).tickSizeOuter(0));
-            
+            .call(axisRight(yScale).tickSizeOuter(0).ticks(5).tickFormat(format('.0f')))
+            .selectAll('.tick')
+            .filter((d) => !Number.isInteger(d))
+            .attr('class', 'hidden');
+
         svgSelection.selectAll('circle')
             .data(data_filtered)
             .join(
@@ -56,7 +61,7 @@
                         .attr('cy', (d) => yScale(d[1]))
                 }
             );
-            
+
             svgSelection.selectAll('.stem')
             .data(data_filtered)
             .join(