+++ src/components/StateSummary.svelte
@@ -1,30 +1,21 @@
 <script>
     import { LeafletMap, StateMap, CumulativeDataTable, FireCountChart, DistrictBarchart } from '$components';
-    import { state_info } from '$lib/data/state_info';
-    import { browser } from '$app/environment';
-    import { getDateISO, IntersectionObserver } from '$lib/utils';
+    import { states } from '$lib/data/site_info.js';
+    import { fires_data } from '../stores/fires_data.js';
+    import { IntersectionObserver } from '$lib/utils';
 
 
 
-    export let state_code, todays_data, district_data, historical_data, layout;
+    export let state_code, todays_data, districts_data, historical_data, layout;
 
-    //TODO try using tanstack-query instead
-    if (browser) {
-        fetch(`https://stub-the-stubble.github.io/data-pipeline/v2/${state_code}/historical_data.json`)
-            .then((res) => res.json())
-            .then((data) => {
-                historical_data = data;
-            });
+    $: {
+        layout = layout ?? 'default';
 
-        //Get date string YYYY-MM-DD format in IST
-        const currentDateStr = getDateISO();
-
-        fetch(`https://stub-the-stubble.github.io/data-pipeline/v2/${state_code}/${currentDateStr}.json`)
-            .then((res) => res.json())
-            .then((data) => {
-                todays_data = data;
-                district_data = Object.entries(todays_data.districts);
-            });
+        if ($fires_data) {
+            todays_data = $fires_data[state_code + '_' + 'today'];
+            districts_data = Object.entries(todays_data.districts);
+            historical_data = $fires_data[state_code + '_' + 'historical'];
+        }
     }
 </script>
 
@@ -33,13 +24,13 @@
 <div>
     <div class="my-12 xs:my-16">
         <h2 class="mb-6 text-5xl text-brown font-bold uppercase">
-            <a href="{state_info[state_code].url}" class="hover:underline">
-                {state_info[state_code].name}
+            <a href="{states[state_code].url}" class="hover:underline">
+                {states[state_code].name}
             </a>
         </h2>
         <CumulativeDataTable data={historical_data?.total.dates} last_updated={todays_data?.last_update} {layout} />
     </div>
-    <div class="flex flex-col {layout === 'narrow' ? '' : 'md:flex-row'} md:items-center gap-x-16">
+    <div class="flex flex-col {layout === 'narrow' ? '' : 'md:flex-row'} md:items-center md:gap-16">
         <IntersectionObserver>
             <div class="mb-8 aspect-w-1 aspect-h-1">
                 <StateMap district_counts={todays_data?.districts} state_code={state_code} />
@@ -51,7 +42,7 @@
                     <h3 class="mb-6 text-xl font-semibold capitalize">
                         Districts with the most stubble fires
                     </h3>
-                    <DistrictBarchart total_count={todays_data.total} {district_data} />
+                    <DistrictBarchart total_count={todays_data.total} {districts_data} />
                 </div>
             {/if}
             <div class="mb-12 last:mb-0">
@@ -63,6 +54,6 @@
         </IntersectionObserver>
     </div>
     <IntersectionObserver>
-        <LeafletMap locations_data={todays_data?.locations} />
+        <!-- <LeafletMap locations_data={todays_data?.locations} /> -->
     </IntersectionObserver>
 </div>